Redo the connector connections and routing
Using 1.5mm pitch to route the connector to the other side of the board
wasn't feasible with the cut copper tape. This expands out the though
holes a bit and increases the pad sizes. It also scootches them to the side
a bit for better trace routing, and adds 1mm to the length of the pcb,
for better printability of the small connector holes near the edge.

This also pushes out the back extension, to give more room for the mounting
standoffs