Strengthen the ballscrew neck
The screws were breaking too easily. They always failed at the joint
between the start of neck and the end of the screws.

This increases the diameter of the neck, so it can bond to the lower
layers better.

This also lengthens the neck, because the upper few threads weren't
actually usable, and were slightly reducing the range of motion of the
joint