Tweak the design of the center key
There's now a single post in the middle of the key, along with a protrusion
to block the light when pressed. In the cluster, there's a central hole
that the post is inserted in to.

The post is also lengthened, so it extends to the bottom of the cluster in
the neutral position, which requires a corresponding hole in the PCB now
too.

And finally the post is enlarged a bit for a smaller clearance with the
central hole in the cluster.

This makes the central key more stable, so it doesn't feel as "wobbly",
and less prone to binding.