Add a rim and profiled post for the center key
This adds an outer rim on the underside of the key, so that the key stops
before where the post meets the key. This helps to prevent binding, since
it can be hard to smooth that area right up against the key bottom.

This also adds a bit of a profile for the post, so that the middle is
slightly smaller. This also helps reduce binding, since only the top and
bottom of the post should touch the post hole now, instead of the entire
length of the post.