Get rid of the peaked magnet hole, and move the center magnets down a bit more
The peaked magnet hole worked, but I didn't particularly like how it looked.
This changes it back to the older squarish magnet hole.

The nubbin on the center key was still interfering with the pliers a bit
when inserting the magnet, so this also moves the magnets down a bit
further.