Increase the y size of the central post by .1
The central hole is square, and since there is no gap between the
front face of the central post and the hole, the gap on the back side is
twice as large as that on the 2 sides.

This increases the size of the post so the gap between the post and hole
on the back and sides is the same.