Widen the key stems for the side keys.
This tweaks the positioning of the optical components to maximize the space
available for the key stems. This will hopefully strengthen the key stems
and help prevent fatique and cracking in the area around the magnet.

That is the area with the least amount of plastic, and it's also where
the top of the key backstop is, so any bending force on the key tends to be
concentrated there.