Insert LEDs/PTs from the bottom
This makes it so that the PCB can be removed from the cluster without
having to unsolder all the LEDs/PTs.

Previously, the pcb was held on via the soldered leads of the LEDs/PTs.
The new design doesn't yet have an attachment mechanism, so that will be
implemented in a later change.

As part of this change, the LED part was switched out to a similar part
that has the same form factor as the phototransistor. This is a slightly
smaller part, so everything was adjusted accordingly.

The height of all of the LEDs/PTs was tweaked slightly, so the top is even
with the top of the cluster, instead of being inset just slightly. This
will make it easy to insert them at the correct height from the bottom.

Additionally, the LED/PT for the central key were raised up to the same
level, instead of being sunken down deeper, and the nub on the central
key adjusted accordingly.