Redo the front part of the cluster
This makes the front part thinner, to put the attachment point higher up,
so there's more room for the screw base, screw and ball underneath. The
front attachment is typically the one closest to the base, so this gives
more space underneath, to allow it to be as close to the base as possible.