Add a clip for the back part of the cluster
The clip fits over the magnets in the back part of the cluster. It serves
a couple of purposes - it provides a spherical divot to keep ball magnets
from sliding, and it has tabs that go over the pcb, to hold it against the
bottom of the cluster and keep it from falling out.