Tweak and simplify the center key for the new cluster design
This shortens the center post, corresponding to the thinner thickness of
the cluster itself.

This also removes a lot of the complexity of the center post. The previous
design was attempting to be clever, only having wide parts in the critical
areas that needed it, in order to reduce the chance of binding (among other
things). But I don't think it's really needed. A straight rectangular post
seems to work reasonably well.