Switch out all the ballscrew designs for the new magnetic screw designs
Instead of having different base sizes, there's just a short base, but
then there are different sized extensions that you can add to get the
desired length.

This also reduces the number of screws to 2. A tiny one for when you need
a really low clearance (i.e. the front of the pinky clusters), and a
somewhat longer one that can be used everywhere else.