Flatten out the bottom of the key well and shorten it by .6mm
This truncates the "V" bottom, so we can shorten the bottom of the key well
a bit.

This will make the outside bottoms of the keywell nearly flush with the
back of the pcb, to enable the use of a 90 degree connector, that would
otherwise intersect the keywell sticking out past the pcb.

This also has some other minor tweaks to ensure horizontal surfaces are
multiples of .2mm away from the print surface (for printing with .2mm
layer height)