Add a virtual "locating pin" to the cluster pcb sketch
This gives a reference point for the origin of the sketch, so the import
location can be set correctly when importing the outline into kicad.

Basically, you set the grid origin in kicad to the center of the circle
(assuming the locating pin is already in kicad) and then make a note of
the coordinates of that point. And then, when importing the outline, you
can specify those coordinates as the location, to ensure the origin of the
imported outline matches that of the existing/previous outline.