Tweak the nut cutouts
On the finger cluster, the nut cutout was 2.6mm tall, but it wasn't aligned
with a .2mm layer line, so it was actually 2.8mm tall once sliced. This
increases the size to 2.8mm and aligns it to a .2mm layer line.

For the thumb clusters, the size was 2.6mm tall and it was already aligned,
but 2.6mm was a bit too short and the nut didn't fit. This increases the
size to 2.8mm.