Update thumb cluster pcbs to v0.2
This shuffles the switch numbering and order on the connector,
so that the center key is on the same pin of the connector (pin 4)
as the center key of the finger clusters.

These keys on the thumb and finger clusters have inverted logic from
all the other keys, due to the physical design of the switches. This
puts the inverted keys on the same pin/column. This simplifies the
firmware a bit, because it can just assume that pin is always inverted.