Tweak the thumb mode keys for better printability
This adds a small "extension" to the beginning of the angled part that
keeps it attached to the bed. The very first part of the slope up would
always merge too strongly with the supports, leaving a bit of support
material on the back.

This gives a bit more vertical room at the point where the supports start,
to help alleviate this problem.