Add a design for a new base, that the steel sheet fits into
This base is 2 plastic pieces that fit together, and that the steel plate
fits into and in-between.

In the previous design, with the handrest attached to a steel plate along
with the clusters, the handrest was prone to sliding around, losing its
place relative to the clusters. This design solves that by firmly
attaching the handrest to magnets in the plastic base, instead of
attaching to the steel plate. This ensures the handrest can't slide
around, relative to the clusters.