Reduce the travel angle of the thumb down key from 9 to 7 degrees
The key was triggering a bit too early/releasing a bit too late, causing
the shift key to accidentally be held down for a key after the intended
key.

This will cause the key to bottom out sooner, so that there should be less
delay from starting to release the key to when it actually triggers the
key up event.