Push the outer thumb keys out just a bit
There wasn't quite enough room between the inner and outer keys, and I
would accidentally rub against the outer keys when flexing my thumb to
press the inner key.

I compared the new thumb cluster with the v1 version, and the inner key
on the v2 is pushed forward just a bit from where it was on the v1. But
it's harder to push back the inner key on v2, with the mode key right
behind it, so I'll try pushing back the outer keys instead.