Change down the thumb down key's travel is limited
This removes the sloped stops on the sides of the hole underneath the down
key. Instead, a post is added to the bottom of the down key at the tip. The
post will hit the pcb and stop the key travel.

I'm not happy with the movement or feel of the thumb down keys. The sloped
stops are a bit problematic, especially if the key is pressed toward the
side, which can cause the edge of the key to slide down the sloped stop a
bit, and skew the whole key in the y axis (the direction the thumb is
pointing).

This should provide a firmer stop, and prevent the key from rotating like
that when pressed on the side.