New center key mechanism
This implements a new center key mechanism that uses a magnet on
the bottom of the cluster, attached to a small nickel strip below
the cluster.

The center key has 2 posts going through the cluster, that attach
to the nickel strip magnetically and when pressed, push it down,
away from the magnet on the bottom of cluster.

This is just the 3d printed parts. The pcb changes will come later.