Add a front magnet to the clusters
Add a magnet, instead of that silly clip thing.

An extra clip on the front isn't normally needed now, but one can
be added in cases where the front of the cluster is too low to the
base plate, to give a bit more room for the support underneath.

This also tweaks the static supports to have another magnet at the
front to attach to the cluster, and moves the cord slot a bit, to
avoid the new magnet.